prototype-backend/requirements.txt
django==5.0.6
django-form-action==1.0.5
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.1
django-cors-headers==4.6.0
django-filter==24.3
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic-settings==2.3.1
requests
pandas==2.2.3
ipython==8.14.0
pyproj==3.7.0
huey==2.4.3
redis==4.3.4
gridstatus==0.28.0
gunicorn==23.0.0
whitenoise==6.8.2
django-timescaledb==0.2.13
gdal==3.10.2